Make Postcode Search a component for reusability
